inventory_service/requirements.txt
flask==2.0.1
werkzeug==2.0.3
flask-graphql==2.0.1
graphene==2.1.9
mysql-connector-python==8.0.27
requests==2.26.0